% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrmaterials-package.R
\docType{package}
\name{mrmaterials-package}
\alias{mrmaterials}
\alias{mrmaterials-package}
\title{mrmaterials: input data generation for material flow analysis models}
\description{
Input data generation for material flow analysis models using the madrat framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/mrmaterials}
}

}
\author{
\strong{Maintainer}: Falk Benke \email{benke@pik-potsdam.de}

Authors:
\itemize{
  \item Jakob Dürrwächter
}

}
\keyword{internal}
